Separate JWT logic into another service
