+++ WarehouseManagement.Api/Extensions/ServiceCollectionExtension.cs
@@ -23,6 +23,7 @@ namespace WarehouseManagement.Api.Extensions
             services.AddScoped<IDeliveryService, DeliveryService>();
             services.AddScoped<IDifferenceTypeService, DifferenceTypeService>();
             services.AddScoped<IDifferenceService, DifferenceService>();
+            services.AddScoped<IJwtService, JwtService>();
             services.AddScoped<IAuthService, AuthService>();
             return services;
         }
