+++ WarehouseManagement.Core/Contracts/IJwtService.cs
@@ -1,0 +1,14 @@
+﻿using WarehouseManagement.Infrastructure.Data.Models;
+
+namespace WarehouseManagement.Core.Contracts;
+
+public interface IJwtService
+{
+    string ComposeAccessToken(string userId, string username, string email);
+
+    Task<string> GenerateRefreshToken(string userId);
+
+    Task<string> GenerateAccessTokenFromRefreshToken(string refreshToken);
+
+    Task RevokeOldRefreshTokens(ApplicationUser user);
+}
